fix: add list-connections command
