+++ applications/tari_base_node/src/commands/parser.rs
@@ -41,7 +41,6 @@ pub enum BaseNodeCommand {
     GetPeer,
     ResetOfflinePeers,
     RewindBlockchain,
-    ListConnections,
     ListHeaders,
     CheckDb,
     PeriodStats,
